requirements.txt
beautifulsoup4
python-dotenv
requests
requests-toolbelt
requests-unixsocket
selenium